Fix 32-bit overflow issue in string table endianness detection

+++ Sources/MachOKit/MachOFile+Strings.swift
@@ -112,7 +112,7 @@ extension MachOFile.UnicodeStrings.Iterator {
         case 2:
             return ptr.pointee == 0xFFFE /* ZERO WIDTH NO-BREAK SPACE (swapped) */
         case 4:
-            return ptr.pointee == 0xFFFE0000
+            return ptr.pointee == UInt32(0xFFFE0000) // avoid overflows in 32bit env
         default:
             return false
         }
